new regression that uses data balancing

+++ effnet_regression_4_balanced_training.py
@@ -394,9 +394,9 @@ class AdamW(Optimizer):
 for cv_index in range(1):
     fold = cv_index
     train_x = train['id_code']
-    train_y = train['diagnosis']
+    train_y = train['diagnosis'].astype(int)
     val_x = val['id_code']
-    val_y = val['diagnosis']
+    val_y = val['diagnosis'].astype(int)
     train_generator = My_Generator(train_x, train_y, batch, is_train=True, augment=True)
     # train_mixup = My_Generator(train_x, train_y, batch, is_train=True, mix=True, augment=True)
     val_generator = My_Generator(val_x, val_y, batch, is_train=False)
